Remove stuff that doesn't work and won't work
Move callbacks into different files

Separate game stuff from memory stuff

Implement required interfaces (WIP)
