+++ src/SourceAutoRecord/Features/Stats.hpp
@@ -3,11 +3,9 @@
 namespace Stats
 {
 	uint32_t TotalJumps;
-	uint32_t TotalUses;
 
 	void Reset()
 	{
 		TotalJumps = 0;
-		TotalUses = 0;
 	}
 }
