+++ src/SourceAutoRecord/Hooks.hpp
@@ -7,11 +7,138 @@
 #include "Modules/Server.hpp"
 
 #include "Game.hpp"
-#include "Utils.hpp"
-
 #include "Interfaces.hpp"
 #include "Patterns.hpp"
+#include "Utils.hpp"
 
+namespace Hooks
+{
+	ScanResult Find(const char* pattern)
+	{
+		auto result = Scan(Patterns::Get(pattern));
+		if (result.Found) {
+			Console::DevMsg("%s\n", result.Message);
+		}
+		else {
+			Console::DevWarning("%s\n", result.Message);
+		}
+		return result;
+	}
+	void Load()
+	{
+		Interfaces::IGameMovement = Interfaces::Get("./portal2/bin/server.so", "GameMovement001");
+		if (Interfaces::IGameMovement) {
+			Console::PrintActive("Found interface IGameMovement (GameMovement001) in server.so!\n");
+			Server::g_GameMovement = std::make_unique<VMTHook>(Interfaces::IGameMovement);
+			Server::g_GameMovement->HookFunction((void*)Server::CheckJumpButton, Offsets::CheckJumpButton);
+		}
+		else {
+			Console::Warning("Failed to get interface IGameMovement (GameMovement001) in server.so!\n");
+		}
+
+		Interfaces::IVEngineClient = Interfaces::Get("./bin/engine.so", "VEngineClient015");
+		if (Interfaces::IVEngineClient) {
+			Console::PrintActive("Found interface IVEngineClient (VEngineClient015) in engine.so!\n");
+			Engine::engine = std::make_unique<VMTHook>(Interfaces::IVEngineClient);
+		}
+		else {
+			Console::Warning("Failed to get interface IVEngineClient (VEngineClient015) in engine.so!\n");
+		}
+
+		auto drecorder = Find("demorecorder");
+		if (drecorder.Found) {
+			Console::PrintActive("Found CDemoRecorder s_ClientDemoRecorder in engine.so!\n");
+			DemoRecorder::Set(drecorder.Address);
+			Engine::s_ClientDemoRecorder = std::make_unique<VMTHook>(**(void***)drecorder.Address);
+			Engine::s_ClientDemoRecorder->HookFunction((void*)Engine::SetSignonState, Offsets::SetSignonState);
+			Engine::s_ClientDemoRecorder->HookFunction((void*)Engine::StopRecording, Offsets::StopRecording);
+		}
+		else {
+			Console::Warning("Failed to get CDemoRecorder s_ClientDemoRecorder in engine.so!\n");
+		}
+
+		auto dplayer = Find("demoplayer");
+		if (drecorder.Found) {
+			DemoPlayer::Set(dplayer.Address);
+			Console::PrintActive("Found CDemoPlayer s_ClientDemoPlayer in engine.so!\n");
+			Engine::s_ClientDemoPlayer = std::make_unique<VMTHook>(**(void***)dplayer.Address);
+			Engine::s_ClientDemoPlayer->HookFunction((void*)Engine::StartPlayback, Offsets::StartPlayback);
+		}
+		else {
+			Console::Warning("Failed to get CDemoPlayer s_ClientDemoPlayer in engine.so!\n");
+		}
+
+		// CClientState cl (function)
+		// \x55\x89\xE5\x83\xEC\x00\xC7\x44\x24\x00\x00\x00\x00\x00\xC7\x04\x24\x00\x00\x00\x00\xE8\x00\x00\x00\x00\xC9\xC3\x8D\x74\x26\x00\x55\x89\xE5\x57\x56\x53\x83\xEC\x00\xA1\x00\x00\x00\x00 xxxxx?xxx?????xxx????x????xxxxxxxxxxxxxx?x????
+		auto cstate = Scan("engine.so", "55 89 E5 83 EC ? C7 44 24 ? ? ? ? ? C7 04 24 ? ? ? ? E8 ? ? ? ? C9 C3 8D 74 26 00 55 89 E5 57 56 53 83 EC ? A1 ? ? ? ? ");
+		if (cstate.Found) {
+			Console::PrintActive("Found CClientState function in engine.so!\n");
+			typedef int (*_GetClientState)();
+			auto GetClientState = reinterpret_cast<_GetClientState>(cstate.Address);
+			Engine::cl = std::make_unique<VMTHook>((void*)GetClientState());
+			Engine::cl->HookFunction((void*)Engine::Disconnect, Offsets::Disconnect);
+		}
+		else {
+			Console::Warning("Failed test in engine.so!\n");
+		}
+
+		Interfaces::IBaseClientDLL = Interfaces::Get("./portal2/bin/client.so", "VClient016");
+		if (Interfaces::IBaseClientDLL) {
+			Console::PrintActive("Found interface IBaseClientDLL (VClient016) in client.so!\n");
+			auto clientdll = std::make_unique<VMTHook>(Interfaces::IBaseClientDLL);
+			auto HudUpdate = clientdll->GetOriginalFunction<uintptr_t>(Offsets::HudUpdate) + 13;
+			Engine::gpGlobals = **reinterpret_cast<CGlobalVarsBase***>(HudUpdate);
+		}
+		else {
+			Console::Warning("Failed to get interface IBaseClientDLL (VClient016) in client.so!\n");
+		}
+
+		auto ldg = Find("m_bLoadgame");
+		auto mpn = Find("m_szMapname");
+		auto ins = Find("g_pInputSystem");
+		auto ksb = Find("Key_SetBinding");
+
+		if (ldg.Found && mpn.Found && ins.Found && ksb.Found) {
+			Console::PrintActive("Found required engine stuff in engine.so!\n");
+			Engine::Set(ldg.Address, mpn.Address);
+			InputSystem::Set(ins.Address, ksb.Address);
+		}
+		else {
+			Console::Warning("Failed to get required engine stuff engine.so!\n");
+		}
+	}
+	bool LoadTier1()
+	{
+		auto cvr = Find("CvarPtr");
+		auto cnv = Find("ConVar_Ctor3");
+		auto cnc = Find("ConCommand_Ctor1");
+		auto cnc2 = Find("ConCommand_Ctor2");
+
+		if (!cvr.Found || !cnv.Found || !cnc.Found || !cnc2.Found)
+			return false;
+
+		Cvar::Set(cvr.Address);
+		Tier1::SetConVar(cnv.Address);
+		Tier1::SetConCommand(cnc.Address, cnc2.Address);
+		return true;
+	}
+	bool LoadClient()
+	{
+		auto sts = Find("SetSize");
+		auto glp = Find("GetLocalPlayer");
+		auto gtp = Find("GetPos");
+		auto mss = Find("g_pMatSystemSurface");
+
+		if (!sts.Found || !glp.Found || !gtp.Found|| !mss.Found)
+			return false;
+
+		Client::Set(sts.Address, glp.Address, gtp.Address);
+		Surface::Set(mss.Address);
+		return true;
+	}
+}
+
+//Interfaces::s_EntityList = GetInterface("./portal2/bin/client.so", "VClientEntityList003");
 /* class IClientEntityList {
 	public:
 		virtual void* GetClientNetworkable(int entindex) = 0;
@@ -24,75 +151,3 @@
 		virtual void SetMaxEntities(int max_entities) = 0;
 		virtual int GetMaxEntities() = 0;
 }; */
-
-namespace Hooks
-{
-	void Load()
-	{
-		Interfaces::g_GameMovement = GetInterface("./portal2/bin/server.so", "GameMovement001");
-		Interfaces::s_EntityList = GetInterface("./portal2/bin/client.so", "VClientEntityList003");
-
-		if (Interfaces::g_GameMovement) {
-			Console::Msg("Found interface IGameMovement (GameMovement001) in server.so\n");
-
-			Server::Hooks::CheckJumpButton = std::make_unique<VMTHook>(Interfaces::g_GameMovement);
-			Server::Hooks::CheckJumpButton
-				->HookFunction((void*)Server::Detour::CheckJumpButton, Offsets::CheckJumpButton);
-		}
-		else
-			Console::Msg("Failed to get interface IGameMovement (GameMovement001) in server.so!\n");
-		
-		if (Interfaces::s_EntityList) {
-			Console::Msg("Found interface IClientEntityList (VClientEntityList003) in client.so\n");
-
-			/* using _GetClientEntity = void*(__cdecl*)(void* thisptr, int entnum);
-			auto GetClientEntity = *(_GetClientEntity*)(*((uintptr_t*)Interfaces::s_EntityList) + 12);
-
-			auto player = GetClientEntity(Interfaces::s_EntityList , 1);
-			if (!player) Console::DevWarning("Failed to find C_BasePlayer!\n");
-			else {
-				Server::Hooks::PlayerUse = std::make_unique<VMTHook>(player);
-				Server::Hooks::PlayerUse
-					->HookFunction((void*)Server::Detour::PlayerUse, Offsets::PlayerUse);
-			} */
-		}
-		else
-			Console::Msg("Failed to get interface IClientEntityList (VClientEntityList003) in client.so!\n");
-
-		// server.so
-		//Create("CheckJumpButton", &Server::Hooks::CheckJumpButton, (void*)Server::Detour::CheckJumpButton);
-		/* Create("PlayerUse", Server::Hooks::PlayerUse, (void*)Server::Detour::PlayerUse);
-
-		// client.so
-		Create("PlayerUse", Client::Hooks::Paint, (void*)Client::Detour::Paint);
-		Create("PlayerUse", Client::Hooks::ShouldDraw, (void*)Client::Detour::ShouldDraw);
-		Create("PlayerUse", Client::Hooks::FindElement, (void*)Client::Detour::FindElement);
-
-		// engine.so
-		Create("SetSignonState", Engine::Hooks::SetSignonState, (void*)Engine::Detour::SetSignonState);
-		Create("CloseDemoFile", Engine::Hooks::SetSignonState, (void*)Engine::Detour::CloseDemoFile);
-		Create("StopRecording", Engine::Hooks::SetSignonState, (void*)Engine::Detour::StopRecording);
-		Create("StartupDemoFile", Engine::Hooks::SetSignonState, (void*)Engine::Detour::StartupDemoFile);
-		Create("ConCommandStop", Engine::Hooks::SetSignonState, (void*)Engine::Detour::ConCommandStop);
-		Create("Disconnect", Engine::Hooks::SetSignonState, (void*)Engine::Detour::Disconnect);
-		Create("PlayDemo", Engine::Hooks::SetSignonState, (void*)Engine::Detour::PlayDemo);
-		Create("StartPlayback", Engine::Hooks::SetSignonState, (void*)Engine::Detour::StartPlayback);
-		Create("HostStateFrame", Engine::Hooks::SetSignonState, (void*)Engine::Detour::HostStateFrame); */
-		
-		/*
-		// Mid-function-hooks
-		auto arm = Scan(Patterns::Get("AirMove"));
-		Server::Original::AirMove = (void*)(arm.Address + 7);
-		JumpToAt((void*)Server::Detour::AirMove, (void*)arm.Address);
-		DoNothingAt(arm.Address + 5, 2);
-		Server::Original::AirMoveSkip = (void*)(arm.Address + Offsets::AirMoveSkip);
-
-		if (Game::Version == Game::Portal2) {
-			auto prc = Scan(Patterns::Get("PlayerRunCommand"));
-			Server::Original::PlayerRunCommand = (void*)(prc.Address + 7);
-			JumpToAt((void*)Server::Detour::PlayerRunCommand, (void*)prc.Address);
-			DoNothingAt(prc.Address + 5, 2);
-			Server::Original::PlayerRunCommandSkip = (void*)(prc.Address + Offsets::PlayerRunCommandSkip);
-		} */
-	}
-}
