+++ src/SourceAutoRecord/Modules/Engine.hpp
@@ -7,100 +7,146 @@
 #include "DemoRecorder.hpp"
 #include "DemoPlayer.hpp"
 
-#include "Demo.hpp"
-#include "Rebinder.hpp"
-#include "Session.hpp"
-#include "Stats.hpp"
-#include "Summary.hpp"
-#include "Timer.hpp"
+#include "Features/Demo.hpp"
+#include "Features/Rebinder.hpp"
+#include "Features/Session.hpp"
+#include "Features/Stats.hpp"
+#include "Features/Summary.hpp"
+#include "Features/Timer.hpp"
+
+#include "Interfaces.hpp"
 #include "Utils.hpp"
 
-using _ClientCmd = void(__fastcall*)(void* thisptr, int edx, const char* szCmdString);
-using _GetGameDir = void(__cdecl*)(char* szGetGameDir, int maxlength);
+using _ClientCmd = int(__cdecl*)(void* thisptr, const char* szCmdString);
+using _GetGameDirectory = char*(__cdecl*)();
 
-using _SetSignonState = bool(__cdecl*)(void* thisptr, int state, int spawncount);
-using _CloseDemoFile = int(__cdecl*)(void* thisptr);
+using _SetSignonState = void(__cdecl*)(void* thisptr, int state);
 using _StopRecording = int(__cdecl*)(void* thisptr);
-using _StartupDemoFile = void(__cdecl*)(void* thisptr);
-using _ConCommandStop = void(__cdecl*)();
-using _Disconnect = void(__cdecl*)(void* thisptr, int bShowMainMenu);
-using _PlayDemo = void(__cdecl*)(void* thisptr);
-using _StartPlayback = bool(__cdecl*)(void* thisptr, const char* filename, bool bAsTimeDemo);
-using _StopPlayback = int(__fastcall*)(void* thisptr, int edx);
-using _HostStateFrame = int(__cdecl*)(float time);
-
-enum SignonState {
-	None = 0,
-	Challenge = 1,
-	Connected = 2,
-	New = 3,
-	Prespawn = 4,
-	Spawn = 5,
-	Full = 6,
-	Changelevel = 7
-};
-
-struct HostStateData {
-	int	m_currentState;
-	int	m_nextState;
-	Vector m_vecLocation;
-	QAngle m_angLocation;
-	char m_levelName[256];
-	char m_landmarkName[256];
-	char m_saveName[256];
-	float m_flShortFrameTime;
-	bool m_activeGame;
-	bool m_bRememberLocation;
-	bool m_bBackgroundLevel;
-	bool m_bWaitingForConnection;
-};
+using _StartPlayback = int(__cdecl*)(void* thisptr, const char* filename, bool bAsTimeDemo);
+using _Disconnect = int(__cdecl*)(void* thisptr, int bShowMainMenu);
 
 using namespace Commands;
 
 namespace Engine
 {
-	void* ClientPtr;
-	_GetGameDir GetGameDir;
-	_ClientCmd ClientCmd;
+	std::unique_ptr<VMTHook> s_ClientDemoRecorder;
+	std::unique_ptr<VMTHook> s_ClientDemoPlayer;
+	std::unique_ptr<VMTHook> engine;
+	std::unique_ptr<VMTHook> cl;
 
-	int* TickCount;
-	float* IntervalPerTick;
-	bool* LoadGame;
-	char** Mapname;
+	CGlobalVarsBase* gpGlobals;
+	
+	int* TickCount = &gpGlobals->tickcount;
+	float* IntervalPerTick = &gpGlobals->interval_per_tick;
 
-	void* CurrentStatePtr;
-
-	void Set(uintptr_t clientPtr, uintptr_t gameDirAddr, uintptr_t curtimeAddr,
-		uintptr_t loadGameAddr, uintptr_t mapnameAddr, uintptr_t currentStateAddr)
-	{
-		ClientPtr = **(void***)(clientPtr);
-		ClientCmd = (_ClientCmd)GetVirtualFunctionByIndex(ClientPtr, Offsets::ClientCommand);
-		GetGameDir = (_GetGameDir)gameDirAddr;
-
-		TickCount = (int*)reinterpret_cast<uintptr_t*>(*((uintptr_t*)curtimeAddr) + Offsets::tickcount);
-		IntervalPerTick = (float*)reinterpret_cast<uintptr_t*>(*((uintptr_t*)curtimeAddr) + Offsets::interval_per_tick);
-		LoadGame = *(bool**)(loadGameAddr);
-		Mapname = (char**)(mapnameAddr);
-
-		CurrentStatePtr = (void*)reinterpret_cast<uintptr_t*>(*((uintptr_t*)currentStateAddr) - 4);
-	}
 	void ExecuteCommand(const char* cmd)
 	{
-		ClientCmd(ClientPtr, 0, cmd);
+		engine->GetOriginalFunction<_ClientCmd>(Offsets::ClientCmd)(Interfaces::IVEngineClient, cmd);
 	}
 	int GetTick()
 	{
-		int result = *TickCount - Session::BaseTick;
+		int result = gpGlobals->tickcount - Session::BaseTick;
 		return (result >= 0) ? result : 0;
 	}
 	std::string GetDir()
 	{
-		char dir[256];
-		GetGameDir(dir, 256);
+		auto dir = engine->GetOriginalFunction<_GetGameDirectory>(Offsets::GetGameDirectory)();
 		return std::string(dir);
 	}
-	void SessionEnded()
+
+	bool* LoadGame;
+	char** Mapname;
+
+	void Set(uintptr_t loadGameAddr, uintptr_t mapnameAddr)
 	{
+		LoadGame = *(bool**)(loadGameAddr);
+		Mapname = (char**)(mapnameAddr);
+	}
+
+	bool IsRecordingDemo;
+	bool IsPlayingDemo;
+
+	void __cdecl SetSignonState(void* thisptr, int state)
+	{
+		Console::PrintActive("SetSignonState!\n");
+		//Console::Print("SetSignonState = %i\n", state);
+		if (state == SignonState::Prespawn) {
+			if (Rebinder::IsSaveBinding || Rebinder::IsReloadBinding) {
+				Rebinder::LastIndexNumber = (IsRecordingDemo)
+					? *DemoRecorder::DemoNumber
+					: Rebinder::LastIndexNumber + 1;
+
+				Rebinder::RebindSave();
+				Rebinder::RebindReload();
+			}
+		}
+		// Demo recorder starts syncing from this tick
+		else if (state == SignonState::Full) {
+			Session::Rebase(*Engine::TickCount);
+			Timer::Rebase(*Engine::TickCount);
+
+			if (*DemoRecorder::Recording) {
+				IsRecordingDemo = true;
+				DemoRecorder::SetCurrentDemo();
+			}
+		}
+		s_ClientDemoRecorder->GetOriginalFunction<_SetSignonState>(Offsets::SetSignonState)(thisptr, state);
+	}
+	int __cdecl StopRecording(void* thisptr)
+	{
+		Console::PrintActive("StopRecording!\n");
+		const int LastDemoNumber = *DemoRecorder::DemoNumber;
+
+		// This function does:
+		// m_bRecording = false
+		// m_nDemoNumber = 0
+		int result = s_ClientDemoRecorder->GetOriginalFunction<_StopRecording>(Offsets::StopRecording)(thisptr);
+
+		if (IsRecordingDemo && sar_autorecord.GetBool()) {
+			*DemoRecorder::DemoNumber = LastDemoNumber;
+
+			// Tell recorder to keep recording
+			if (*LoadGame) {
+				*DemoRecorder::Recording = true;
+				(*DemoRecorder::DemoNumber)++;
+			}
+		}
+		else {
+			IsRecordingDemo = false;
+		}
+
+		if (IsPlayingDemo) {
+			IsPlayingDemo = false;
+		}
+
+		return result;
+	}
+	int __cdecl StartPlayback(void* thisptr, const char *filename, bool bAsTimeDemo)
+	{
+		Console::PrintActive("StartPlayback!\n");
+		int result = s_ClientDemoPlayer->GetOriginalFunction<_StartPlayback>(Offsets::StartPlayback)(thisptr, filename, bAsTimeDemo);
+
+		if (result) {
+			IsPlayingDemo = true;
+
+			Demo demo;
+			if (demo.Parse(GetDir() + std::string("\\") + std::string(DemoPlayer::DemoName))) {
+				demo.Fix();
+				Console::Print("Client: %s\n", demo.clientName);
+				Console::Print("Map: %s\n", demo.mapName);
+				Console::Print("Ticks: %i\n", demo.playbackTicks);
+				Console::Print("Time: %.3f\n", demo.playbackTime);
+				Console::Print("IpT: %.6f\n", demo.IntervalPerTick());
+			}
+			else {
+				Console::Print("Could not parse \"%s\"!\n", DemoPlayer::DemoName);
+			}
+		}
+		return result;
+	}
+	int __cdecl Disconnect(void* thisptr, bool bShowMainMenu)
+	{
+		Console::PrintActive("Disconnect!\n");
 		if (!*LoadGame && !DemoPlayer::IsPlaying()) {
 			int tick = GetTick();
 
@@ -131,168 +177,6 @@ namespace Engine
 
 			DemoRecorder::CurrentDemo = "";
 		}
-	}
-
-	namespace Hooks
-	{
-		std::unique_ptr<VMTHook> SetSignonState;
-		std::unique_ptr<VMTHook> CloseDemoFile;
-		std::unique_ptr<VMTHook> StopRecording;
-		std::unique_ptr<VMTHook> StartupDemoFile;
-		std::unique_ptr<VMTHook> ConCommandStop;
-		std::unique_ptr<VMTHook> Disconnect;
-		std::unique_ptr<VMTHook> PlayDemo;
-		std::unique_ptr<VMTHook> StartPlayback;
-		std::unique_ptr<VMTHook> HostStateFrame;
-	}
-
-	namespace Detour
-	{
-		bool PlayerRequestedStop;
-		bool IsRecordingDemo;
-
-		bool PlayerRequestedPlayback;
-		bool IsPlayingDemo;
-
-		int LastHostState;
-		bool CallFromHostStateFrame;
-
-		/* bool __fastcall SetSignonState(void* thisptr, int edx, int state, int spawncount)
-		{
-			//Console::Print("SetSignonState = %i\n", state);
-			if (state == SignonState::Prespawn) {
-				if (Rebinder::IsSaveBinding || Rebinder::IsReloadBinding) {
-					Rebinder::LastIndexNumber = (IsRecordingDemo)
-						? *DemoRecorder::DemoNumber
-						: Rebinder::LastIndexNumber + 1;
-
-					Rebinder::RebindSave();
-					Rebinder::RebindReload();
-				}
-			}
-			// Demo recorder starts syncing from this tick
-			else if (state == SignonState::Full) {
-				Session::Rebase(*Engine::TickCount);
-				Timer::Rebase(*Engine::TickCount);
-			}
-			return ((_SetSignonState)subhook_get_trampoline(Hooks::SetSignonState))(thisptr, state, spawncount);
-		}
-		int __fastcall CloseDemoFile(void* thisptr, int edx)
-		{
-			bool* m_bIsDemoHeader = (bool*)((uintptr_t)thisptr + Offsets::m_bIsDemoHeader);
-
-			if (!*m_bIsDemoHeader && !*LoadGame && IsRecordingDemo) {
-				int tick = DemoRecorder::GetTick();
-				float time = tick * *IntervalPerTick;
-				Console::Print("Demo: %i (%.3f)\n", tick, time);
-				DemoRecorder::SetLastDemo(tick);
-			}
-
-			return ((_CloseDemoFile)subhook_get_trampoline(Hooks::CloseDemoFile))(thisptr);
-		}
-		bool __fastcall StopRecording(void* thisptr, int edx)
-		{
-			const int LastDemoNumber = *DemoRecorder::DemoNumber;
-
-			// This function does:
-			// m_bRecording = false
-			// m_nDemoNumber = 0
-			bool result = ((_StopRecording)subhook_get_trampoline(Hooks::StopRecording))(thisptr);
-
-			if (IsRecordingDemo && !PlayerRequestedStop) {
-				*DemoRecorder::DemoNumber = LastDemoNumber;
-
-				// Tell recorder to keep recording
-				if (*LoadGame || CallFromHostStateFrame) {
-					*DemoRecorder::Recording = true;
-					(*DemoRecorder::DemoNumber)++;
-				}
-			}
-			else {
-				IsRecordingDemo = false;
-			}
-
-			if (IsPlayingDemo && !PlayerRequestedPlayback) {
-				IsPlayingDemo = false;
-			}
-
-			return result;
-		}
-		void __fastcall StartupDemoFile(void* thisptr, int edx)
-		{
-			IsRecordingDemo = true;
-			DemoRecorder::SetCurrentDemo();
-			((_StartupDemoFile)subhook_get_trampoline(Hooks::StartupDemoFile))(thisptr);
-		}
-		void __cdecl ConCommandStop()
-		{
-			PlayerRequestedStop = true;
-			((_ConCommandStop)subhook_get_trampoline(Hooks::ConCommandStop))();
-			PlayerRequestedStop = false;
-		}
-		void __fastcall Disconnect(void* thisptr, int edx, bool bShowMainMenu)
-		{
-			SessionEnded();
-			((_Disconnect)subhook_get_trampoline(Hooks::Disconnect))(thisptr, bShowMainMenu);
-		}
-		void __cdecl PlayDemo(void* thisptr)
-		{
-			PlayerRequestedPlayback = true;
-			((_PlayDemo)subhook_get_trampoline(Hooks::PlayDemo))(thisptr);
-			PlayerRequestedPlayback = false;
-		}
-		bool __fastcall StartPlayback(void* thisptr, int edx, const char *filename, bool bAsTimeDemo)
-		{
-			bool result = ((_StartPlayback)subhook_get_trampoline(Hooks::StartPlayback))(thisptr, filename, bAsTimeDemo);
-
-			if (result && PlayerRequestedPlayback) {
-				IsPlayingDemo = true;
-
-				// Allows sar_time_demo to parse last played demo
-				DemoRecorder::LastDemo == "";
-
-				Demo demo;
-				if (demo.Parse(GetDir() + std::string("\\") + std::string(DemoPlayer::DemoName))) {
-					demo.Fix();
-					Console::Print("Client: %s\n", demo.clientName);
-					Console::Print("Map: %s\n", demo.mapName);
-					Console::Print("Ticks: %i\n", demo.playbackTicks);
-					Console::Print("Time: %.3f\n", demo.playbackTime);
-					Console::Print("IpT: %.6f\n", demo.IntervalPerTick());
-				}
-				else {
-					Console::Print("Could not parse \"%s\"!\n", DemoPlayer::DemoName);
-				}
-			}
-			return result;
-		}
-		int __fastcall HostStateFrame(void* thisptr, int edx, float time)
-		{
-			HostStateData state = *reinterpret_cast<HostStateData*>(CurrentStatePtr);
-
-			if (state.m_currentState != LastHostState) {
-				//Console::Print("m_currentState = %i\n", state.m_currentState);
-				if (state.m_currentState == Offsets::HS_CHANGE_LEVEL_SP) {
-					SessionEnded();
-
-					// CloseDemoFile gets called too late when changing the level
-					// which causes to write invalid ticks into the demo header and redundant packets
-					if (IsRecordingDemo) {
-						CallFromHostStateFrame = true;
-						StopRecording(DemoRecorder::Ptr, 0);
-						CallFromHostStateFrame = false;
-					}
-				}
-
-				// Start new session when in menu
-				if (state.m_currentState == Offsets::HS_RUN && !state.m_activeGame && !DemoPlayer::IsPlaying()) {
-					//Console::Print("Detected menu!\n");
-					Session::Rebase(*Engine::TickCount);
-				}
-
-				LastHostState = state.m_currentState;
-			}
-			return ((_HostStateFrame)subhook_get_trampoline(Hooks::HostStateFrame))(time);
-		} */
+		return cl->GetOriginalFunction<_Disconnect>(Offsets::Disconnect)(thisptr, bShowMainMenu);
 	}
 }
