+++ src/SourceAutoRecord/SourceAutoRecord.hpp
@@ -1,426 +1,6 @@
 #pragma once
-
-#include "Modules/Client.hpp"
-#include "Modules/Console.hpp"
-#include "Modules/DemoPlayer.hpp"
-#include "Modules/DemoRecorder.hpp"
-#include "Modules/Engine.hpp"
-#include "Modules/InputSystem.hpp"
-#include "Modules/Server.hpp"
-
-#include "Modules/ConCommand.hpp"
-#include "Modules/ConVar.hpp"
-#include "Modules/Cvar.hpp"
-
-#include "Callbacks.hpp"
-#include "Commands.hpp"
-#include "Game.hpp"
-#include "Hooks.hpp"
-#include "Offsets.hpp"
-#include "Utils.hpp"
-
-#include "Patterns.hpp"
-
-using namespace Commands;
-using namespace Cvar;
-
-namespace SAR
-{
-	ScanResult Find(const char* pattern)
-	{
-		auto result = Scan(Patterns::Get(pattern));
-		if (result.Found) {
-			Console::DevMsg("%s\n", result.Message);
-		}
-		else {
-			Console::DevWarning("%s\n", result.Message);
-		}
-		return result;
-	}
-	bool LoadTier1()
-	{
-		auto cvr = Find("CvarPtr");
-		auto cnv = Find("ConVar_Ctor3");
-		auto cnc = Find("ConCommand_Ctor1");
-		auto cnc2 = Find("ConCommand_Ctor2");
-
-		if (!cvr.Found || !cnv.Found || !cnc.Found || !cnc2.Found)
-			return false;
-
-		Cvar::Set(cvr.Address);
-		Tier1::SetConVar(cnv.Address);
-		Tier1::SetConCommand(cnc.Address, cnc2.Address);
-		return true;
-	}
-	bool LoadClient()
-	{
-		auto sts = Find("SetSize");
-		auto glp = Find("GetLocalPlayer");
-		auto gtp = Find("GetPos");
-		auto mss = Find("g_pMatSystemSurface");
-
-		if (!sts.Found || !glp.Found || !gtp.Found|| !mss.Found)
-			return false;
-
-		Client::Set(sts.Address, glp.Address, gtp.Address);
-		Surface::Set(mss.Address);
-		return true;
-	}
-	bool LoadEngine()
-	{
-		auto enc = Find("engineClient");
-		auto ggd = Find("GetGameDir");
-		auto crt = Find("curtime");
-		auto ldg = Find("m_bLoadgame");
-		auto mpn = Find("m_szMapname");
-		auto cns = Find("m_currentState");
-		auto drc = Find("demorecorder");
-		auto ins = Find("g_pInputSystem");
-		auto ksb = Find("Key_SetBinding");
-		auto dpl = Find("demoplayer");
-
-		if (!enc.Found || !ggd.Found || !crt.Found || !ldg.Found || !mpn.Found
-			|| !cns.Found || !drc.Found || !ins.Found || !ksb.Found || !dpl.Found)
-			return false;
-
-		Engine::Set(enc.Address, ggd.Address, crt.Address, ldg.Address, mpn.Address, cns.Address);
-		DemoRecorder::Set(drc.Address);
-		InputSystem::Set(ins.Address, ksb.Address);
-		DemoPlayer::Set(dpl.Address);
-		return true;
-	}
-	void CreateCommands()
-	{
-		// Rebinder
-		sar_bind_save = CreateCommandArgs(
-			"sar_bind_save", Callbacks::BindSaveRebinder,
-			"Automatic save rebinding when server has loaded. File indexing will be synced when recording demos. Usage: sar_bind_save <key> [save_name]\n");
-		sar_bind_reload = CreateCommandArgs(
-			"sar_bind_reload", Callbacks::BindReloadRebinder,
-			"Automatic save-reload rebinding when server has loaded. File indexing will be synced when recording demos. Usage: sar_bind_reload <key> [save_name]\n");
-		sar_unbind_save = CreateCommand(
-			"sar_unbind_save", Callbacks::UnbindSaveRebinder,
-			"Unbinds current save rebinder.\n");
-		sar_unbind_reload = CreateCommand(
-			"sar_unbind_reload", Callbacks::UnbindReloadRebinder,
-			"Unbinds current save-reload rebinder.\n");
-		sar_save_flag = CreateString(
-			"sar_save_flag", "#SAVE#",
-			"Echo message when using sar_bind_save. Default is \"#SAVE#\", a SourceRuns standard. Keep this empty if no echo message should be binded.\n");
-
-		// Demo parsing
-		sar_time_demo = CreateCommandArgs(
-			"sar_time_demo",
-			Callbacks::PrintDemoInfo,
-			"Parses a demo and prints some information about it.\n");
-		sar_time_demos = CreateCommandArgs(
-			"sar_time_demos",
-			Callbacks::PrintDemoInfos,
-			"Parses multiple demos and prints the total sum of them.\n");
-		sar_time_demo_dev = CreateFloat(
-			"sar_time_demo_dev",
-			"0",
-			0,
-			"Printing mode when using sar_time_demo. 0 = default, 1 = console commands, 2 = console commands & packets.\n");
-
-		// Summary
-		sar_sum_here = CreateCommand(
-			"sar_sum_here",
-			Callbacks::StartSummary,
-			"Starts counting total ticks of sessions.\n");
-		sar_sum_stop = CreateCommand(
-			"sar_sum_stop",
-			Callbacks::StopSummary,
-			"Stops summary counter.\n");
-		sar_sum_result = CreateCommand(
-			"sar_sum_result",
-			Callbacks::PrintSummary,
-			"Prints result of summary.\n");
-		sar_sum_during_session = CreateBoolean(
-			"sar_sum_during_session",
-			"1",
-			"Updates the summary counter automatically during a session.\n");
-
-		// Timer
-		sar_timer_start = CreateCommand(
-			"sar_timer_start",
-			Callbacks::StartTimer,
-			"Starts timer.\n");
-		sar_timer_stop = CreateCommand(
-			"sar_timer_stop",
-			Callbacks::StopTimer,
-			"Stops timer.\n");
-		sar_timer_result = CreateCommand(
-			"sar_timer_result",
-			Callbacks::PrintTimer,
-			"Prints result of timer.\n");
-		sar_timer_always_running = CreateBoolean(
-			"sar_timer_always_running",
-			"1",
-			"Timer will save current value when disconnecting.\n");
-
-		// Timer average
-		sar_avg_start = CreateCommand(
-			"sar_avg_start",
-			Callbacks::StartAverage,
-			"Starts calculating the average when using timer.\n");
-		sar_avg_stop = CreateCommand(
-			"sar_avg_stop",
-			Callbacks::StopAverage,
-			"Stops average calculation.\n");
-		sar_avg_result = CreateCommand(
-			"sar_avg_result",
-			Callbacks::PrintAverage,
-			"Prints result of average.\n");
-
-		// Timer checkpoints
-		sar_cps_add = CreateCommand(
-			"sar_cps_add",
-			Callbacks::AddCheckpoint,
-			"Saves current time of timer.\n");
-		sar_cps_clear = CreateCommand(
-			"sar_cps_clear",
-			Callbacks::ClearCheckpoints,
-			"Resets saved times of timer.\n");
-		sar_cps_result = CreateCommand(
-			"sar_cps_result",
-			Callbacks::PrintCheckpoints,
-			"Prints result of timer checkpoints.\n");
-
-		// HUD
-		sar_hud_text = CreateString(
-			"sar_hud_text",
-			"",
-			"Draws specified text when not empty.\n");
-		sar_hud_position = CreateFloat(
-			"sar_hud_position",
-			"0",
-			0,
-			"Draws absolute position of the client.\n");
-		sar_hud_angles = CreateBoolean(
-			"sar_hud_angles",
-			"0",
-			"Draws absolute view angles of the client.\n");
-		sar_hud_velocity = CreateFloat(
-			"sar_hud_velocity",
-			"0",
-			0,
-			"Draws velocity of the client. 0 = default, 1 = x/y/z , 2 = x/y\n");
-		sar_hud_session = CreateBoolean(
-			"sar_hud_session",
-			"0",
-			"Draws current session value.\n");
-		sar_hud_last_session = CreateBoolean(
-			"sar_hud_last_session",
-			"0",
-			"Draws value of latest completed session.\n");
-		sar_hud_sum = CreateBoolean(
-			"sar_hud_sum",
-			"0",
-			"Draws summary value of sessions.\n");
-		sar_hud_timer = CreateBoolean(
-			"sar_hud_timer",
-			"0",
-			"Draws current value of timer.\n");
-		sar_hud_avg = CreateBoolean(
-			"sar_hud_avg",
-			"0",
-			"Draws calculated average of timer.\n");
-		sar_hud_cps = CreateBoolean(
-			"sar_hud_cps",
-			"0",
-			"Draws latest checkpoint of timer.\n");
-		sar_hud_demo = CreateBoolean(
-			"sar_hud_demo",
-			"0",
-			"Draws name, tick and time of current demo.\n");
-		sar_hud_last_demo = CreateBoolean(
-			"sar_hud_last_demo",
-			"0",
-			"Draws value of latest completed demo.\n");
-		sar_hud_jumps = CreateBoolean(
-			"sar_hud_jumps",
-			"0",
-			"Draws total jump count.\n");
-		sar_hud_uses = CreateBoolean(
-			"sar_hud_uses",
-			"0",
-			"Draws total use count.\n");
-		sar_hud_default_spacing = CreateFloat(
-			"sar_hud_default_spacing",
-			"4",
-			0,
-			"Spacing between elements of HUD.\n");
-		sar_hud_default_padding_x = CreateFloat(
-			"sar_hud_default_padding_x",
-			"2",
-			0,
-			"X padding of HUD.\n");
-		sar_hud_default_padding_y = CreateFloat(
-			"sar_hud_default_padding_y",
-			"2",
-			0,
-			"Y padding of HUD.\n");
-		sar_hud_default_font_size = CreateFloat(
-			"sar_hud_default_font_size",
-			"10",
-			0,
-			"Font size of HUD.\n");
-		sar_hud_default_font_index = CreateFloat(
-			"sar_hud_default_font_index",
-			"0",
-			0,
-			225,
-			"Font value of HUD.\n");
-		sar_hud_default_font_color = CreateString(
-			"sar_hud_default_font_color",
-			"255 255 255 255",
-			"RGBA font color of HUD.\n");
-
-		// Stats
-		sar_stats_auto_reset = CreateFloat(
-			"sar_stats_auto_reset",
-			"0",
-			0,
-			"Resets all stats automatically. 0 = default, 1 = disconnect, 2 = disconnect & sar_timer_start.\n");
-		sar_stats_reset_jumps = CreateCommand(
-			"sar_stats_reset_jumps",
-			Callbacks::ResetJumps,
-			"Resets jump counter.\n");
-		sar_stats_reset_uses = CreateCommand(
-			"sar_stats_reset_uses",
-			Callbacks::ResetUses,
-			"Resets use counter.\n");
-
-		// Cheats
-		sar_autojump = CreateBoolean(
-			"sar_autojump",
-			"0",
-			"Enables automatic jumping on the server.\n");
-		sar_aircontrol = CreateBoolean(
-			"sar_aircontrol",
-			"0",
-			"Enables \"more air-control\" on the server.\n");
-		if (Game::Version == Game::Portal2) {
-			sar_never_open_cm_hud = CreateBoolean(
-				"sar_never_open_cm_hud",
-				"0",
-				"Removes the challenge mode scoreboard panel.\n");
-		}
-		sar_never_delay_start = CreateBoolean(
-			"sar_never_delay_start",
-			"0",
-			"Doesn't prevent player to be frozen at spawn.\n");
-		sar_teleport = CreateCommand(
-			"sar_teleport",
-			Callbacks::Teleport,
-			"Teleports the player to the last saved location.\n");
-		sar_teleport_setpos = CreateCommand(
-			"sar_teleport_setpos",
-			Callbacks::SetTeleport,
-			"Saves current location for teleportation.\n");
-
-		// Others
-		sar_session = CreateCommand(
-			"sar_session",
-			Callbacks::PrintSession,
-			"Prints the current tick of the server since it has loaded.\n");
-		sar_about = CreateCommand(
-			"sar_about",
-			Callbacks::PrintAbout,
-			"Prints info about this plugin.\n");
-		sar_cvars_save = CreateCommand(
-			"sar_cvars_save",
-			Callbacks::SaveCvars,
-			"Saves important SAR cvars.\n");
-		sar_cvars_load = CreateCommand(
-			"sar_cvars_load",
-			Callbacks::LoadCvars,
-			"Loads important SAR cvars.\n");
-
-		// From the game
-		cl_showpos = FindCvar("cl_showpos");
-		cl_sidespeed = FindCvar("cl_sidespeed");
-		sv_cheats = FindCvar("sv_cheats");
-
-		sv_bonus_challenge = FindCvar("sv_bonus_challenge");
-		sv_accelerate = FindCvar("sv_accelerate");
-		sv_airaccelerate = FindCvar("sv_airaccelerate");
-		sv_friction = FindCvar("sv_friction");
-		sv_maxspeed = FindCvar("sv_maxspeed");
-		sv_stopspeed = FindCvar("sv_stopspeed");
-		sv_maxvelocity = FindCvar("sv_maxvelocity");
-		sv_transition_fade_time = FindCvar("sv_transition_fade_time");
-		sv_laser_cube_autoaim = FindCvar("sv_laser_cube_autoaim");
-		ui_loadingscreen_transition_time = FindCvar("ui_loadingscreen_transition_time");
-		hide_gun_when_holding = FindCvar("hide_gun_when_holding");
-
-		Console::DevMsg("SAR: Created %i ConVars and %i ConCommands!\n", Tier1::ConVarCount, Tier1::ConCommandCount);
-	}
-	void Enable(ConVar& var, bool asCheat = true)
-	{
-		var.RemoveFlag(FCVAR_DEVELOPMENTONLY | FCVAR_HIDDEN);
-		if (asCheat) var.AddFlag(FCVAR_CHEAT);
-	}
-	void EnableGameCheats()
-	{
-		// Challenge mode will reset every cheat automatically
-		// Flagging this as a cheat would break cm. I think
-		// it's impossible to abuse this anyway
-		Enable(sv_bonus_challenge, false);
-
-		Enable(sv_accelerate);
-		Enable(sv_airaccelerate);
-		Enable(sv_friction);
-		Enable(sv_maxspeed);
-		Enable(sv_stopspeed);
-		Enable(sv_maxvelocity);
-
-		if (Game::Version == Game::Portal2) {
-			Enable(sv_transition_fade_time);
-			Enable(sv_laser_cube_autoaim);
-			Enable(ui_loadingscreen_transition_time);
-
-			// Not a real cheat, right?
-			Enable(hide_gun_when_holding, false);
-		}
-	}
-	void LoadPatches()
-	{
-		// Change limited help string output in ConVar_PrintDescription
-		// from 80 to 1024
-		const char* thanksForNothingValve = "%-80s - %.1024s\n";
-		auto prd = Find("PrintDescription");
-		if (prd.Address) {
-			auto result = NewReferenceAt((void*)prd.Address, (void*)thanksForNothingValve);
-			if (result != -1)
-				Console::DevMsg("SAR: Patched ConVar_PrintDescription at 0x%p!\n", prd.Address);
-			else {
-				Console::DevWarning("SAR: Failed to patch ConVar_PrintDescription (%s)!\n", strerror(errno));
-			}
-		}
-
-		/* // Remove Cbuf_AddText and Cbuf_Execute when playing a demo
-		auto rdp = Find("ReadPacket");
-		if (rdp.Found) {
-			DoNothingAt(rdp.Address, 25);
-			Console::DevMsg("SAR: Patched CDemoPlayer::ReadPacket at 0x%p!\n", rdp.Address);
-		}
-
-		if (Game::Version == Game::Portal2) {
-			// Remove the default ConMsg when demo file gets closed
-			auto cdf = Scan("engine.so", "55 89 E5 56 53 83 EC ? 8B 5D ? 8D 73 ? 89 34 24", 284);
-			if (cdf.Found) {
-				DoNothingAt(cdf.Address, 44);
-				Console::DevMsg("SAR: Patched CDemoPlayer::CloseDemoFile at 0x%p!\n", cdf.Address);
-			}
-
-			// Remove automatic -remote_view on map load
-			auto rv = Scan("server.so", "55 89 E5 56 53 83 EC ? 8B 75 ? 89 34 24 E8 ? ? ? ? C7 44 24 ? ? ? ? ? ", 108);
-			if (rv.Found) {
-				DoNothingAt(rv.Address, 32);
-				Console::DevMsg("SAR: Patched -remote_view at 0x%p!\n", rv.Address);
-			}
-		} */
-	}
-}
+#define SAR_VERSION "1.5"
+#define SAR_BUILD __TIME__ " " __DATE__
+#define SAR_COLOR Color(247, 235, 69)
+#define COL_ACTIVE Color(110, 247, 76)
+#define COL_DEFAULT Color(255, 255, 255, 255)
