+++ src/SourceAutoRecord/Utils/SDK.hpp
@@ -1,0 +1,83 @@
+#pragma once
+#include <cmath>
+
+struct Vector {
+	float x, y, z;
+	float Length() {
+		return sqrt(x * x + y * y + z * z);
+	}
+	float Length2D() {
+		return sqrt(x * x + y * y);
+	}
+};
+
+struct QAngle {
+	float x, y, z;
+};
+
+struct Color {
+	Color() {
+		*((int *)this) = 255;
+	}
+	Color(int _r, int _g, int _b) {
+		SetColor(_r, _g, _b, 255);
+	}
+	Color(int _r, int _g, int _b, int _a) {
+		SetColor(_r, _g, _b, _a);
+	}
+	void SetColor(int _r, int _g, int _b, int _a = 255) {
+		_color[0] = (unsigned char)_r;
+		_color[1] = (unsigned char)_g;
+		_color[2] = (unsigned char)_b;
+		_color[3] = (unsigned char)_a;
+	}
+	inline int r() const { return _color[0]; }
+	inline int g() const { return _color[1]; }
+	inline int b() const { return _color[2]; }
+	inline int a() const { return _color[3]; }
+	unsigned char _color[4];
+};
+
+enum SignonState {
+	None = 0,
+	Challenge = 1,
+	Connected = 2,
+	New = 3,
+	Prespawn = 4,
+	Spawn = 5,
+	Full = 6,
+	Changelevel = 7
+};
+
+/* struct HostStateData {
+	int	m_currentState;
+	int	m_nextState;
+	Vector m_vecLocation;
+	QAngle m_angLocation;
+	char m_levelName[256];
+	char m_landmarkName[256];
+	char m_saveName[256];
+	float m_flShortFrameTime;
+	bool m_activeGame;
+	bool m_bRememberLocation;
+	bool m_bBackgroundLevel;
+	bool m_bWaitingForConnection;
+}; */
+
+struct CGlobalVarsBase {
+	float realtime;
+	int framecount;
+	float absoluteframetime;
+	float curtime;
+	float frametime;
+	int maxClients;
+	int tickcount;
+	float interval_per_tick;
+	float interpolation_amount;
+	int simTicksThisFrame;
+	int network_protocol;
+	int* pSaveData;
+	bool m_bClient;
+	int nTimestampNetworkingBase;
+	int nTimestampRandomizeWindow;
+};
